agri-ai/backend/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
pymongo==4.7.2
python-dotenv==1.0.1
requests==2.32.3
scikit-learn==1.5.0
numpy==1.26.4
pandas==2.2.2
gunicorn==22.0.0
dnspython==2.6.1
marshmallow==3.21.3
